move iterator_read_into_vector to common
